Added -h and -no_anims command line args. 64 bit fix: SYS_SOUNDHANDLE is now an unsigned long for 64 bit.

+++ src/system/system.h
@@ -128,7 +128,7 @@ int SYS_KeyPressed( unsigned char key );
 
 
 // sound stuff
-typedef unsigned int SYS_SOUNDHANDLE;
+typedef unsigned long SYS_SOUNDHANDLE;
 void SYS_InitSound();
 void SYS_ShutdownSound();
 void SYS_ProcessSound();	// TODO: maybe this should be hidden from game client.
